cleanup: move json code to Extism.JSON module
