+++ haskell/manifest/extism-manifest.cabal
@@ -11,7 +11,7 @@ category:           Plugins, WebAssembly
 extra-source-files: CHANGELOG.md
 
 library
-    exposed-modules:    Extism.Manifest
+    exposed-modules:    Extism.Manifest Extism.JSON
     hs-source-dirs:     .
     default-language:   Haskell2010
     build-depends:
